:adhesive_bandage: (sonarcloud): Fix bugs and code smells
:adhesive_bandage: (otm8009a): Use std::byte for bitwise operation
:adhesive_bandage: (jpeg): Remove redundant cast
:adhesive_bandage: (dma2d): Fix implicit conversion loses
:adhesive_bandage: (jpeg): Add tests recommanded by sonarcloud
:adhesive_bandage: (hal): Remove unused acces specifier
:adhesive_bandage: (hal): Use =default as default implementation
:adhesive_bandage: const non-mutable member function
:adhesive_bandage: (ll): Remove unused namespace
:adhesive_bandage: (dsi): Avoid implicit conversion
:adhesive_bandage: (lcd) Add missing final keyword
:adhesive_bandage: (font): Add const in fontGetPixelBytes argument
:adhesive_bandage: (otm8009a): Replace "/*" by "//"
:adhesive_bandage: (otm8009a): Nested namespace

+++ drivers/LKCoreVideo/include/internal/LKCoreDMA2D.h
@@ -16,10 +16,10 @@ class LKCoreDMA2D : public LKCoreDMA2DBase
 
 	void initialize() final;
 
-	void transferData(uintptr_t input_data_address, uintptr_t output_data_address, uint16_t width,
-					  uint16_t height) final;
-	void transferImage(uint16_t width, uint16_t height, uint16_t width_offset) final;
-	void transferDrawing(uintptr_t first_pixel_address, uint16_t width, uint16_t height, uint32_t color) final;
+	void transferData(uintptr_t input_data_address, uintptr_t output_data_address, uint32_t width,
+					  uint32_t height) final;
+	void transferImage(uint32_t width, uint32_t height, uint32_t width_offset) final;
+	void transferDrawing(uintptr_t first_pixel_address, uint32_t width, uint32_t height, uint32_t color) final;
 
 	DMA2D_HandleTypeDef getHandle(void) final;
